Create html for to do task input dialog
When the user wants to add a new to do task by clicking the Add Task
button, a dialog needs to appear for them to input the information.
This html element has been added, but not linked to a button yet.
This will be done in a new handler module.