Show new task modal on button click
The modal that allows the user to enter a new task is now opened
when the New Task button is clicked and closed if either the Add Task
button is clicked when the form is filled out or if Cancel is clicked.
The webpage does not do anything with the new task information yet,
such as displaying it. This will be done with a dom content handler
module.