Send new task input to to do handler
When the user clicks Add Task on the new task dialog, the information
needs to be sent to the to do handler to be properly stored and dealt
with. This is done by importing toDos into the listeners module and
calling toDos.addTask when the Add Task button is clicked and all the
fields are properly filled out. For now, the new task is just logged
in the console to ensure every module is communicating properly, but in
the future a dom handler will be called to display the task on the
webpage.