Change visibility of New Task button
The user should only be able to add a new task to a specific project.
Thus, when a new project is created and added to the webpage, the New
Task button is set to visible when that project is clicked. Also, a
new function showAllProjects has been added to the dom handler to hide
the New Task button when all projects are displayed.