Add showAllProjects functionality
When the user clicks the All Tasks button, the display needs to show
all tasks from every project. This is now done in the showAllProjects
method of the dom handler module. The method removes current display
items and iterates through the project list and sub-arrays to display
all project tasks.