Prevent duplicate task names
The task names within each project will serve as a unique identifier
for each task. This will be important when it comes to deleting and
editing tasks. Therefore, a task listener for the task name input has
been added to ensure that there are no duplicate task names within a
single project.