Display All Tasks ordered by due date
Each project displays the tasks in the order in which they were
created by the user. However, the All Tasks section displays all
tasks ordered by due date, soonest first. This is done with the help
of the stored array in the toDos module that already contains all
tasks in sorted order.