Style project sidebar
The sidebar allow the user to view the list of project names and select
a specific project view or choose to view all projects. This sidebar has
been styled, including coloring and animations to make it clear what
project or if all tasks has been selected as well as spacing and font
size changes to make it more readable.